Add semaphore to Wire Protocol
- Add binary sempahore to Wire Protocol to lock access to the tx/rx channel (on heavy message trafic not having this was causing constant buffer overrun and garbled communication)
- Moved WireProtocol_HAL_Interface to target common (this file was being repeated without a single change on every reference board)
