+++ targets/CMSIS-OS/ChibiOS/common/WireProtocol_ReceiverThread.c
@@ -9,6 +9,7 @@
 #include "WireProtocol_HAL_Interface.h"
 
 extern WP_Message inboundMessage;
+extern binary_semaphore_t wpChannelSemaphore;
 
 // This thread needs to be implemented at ChibiOS level because it has to include a call to chThdShouldTerminateX()
 // in case the thread is requested to terminate by the CMSIS call osThreadTerminate()
@@ -16,6 +17,9 @@ extern WP_Message inboundMessage;
 void ReceiverThread(void const * argument)
 {
   (void)argument;
+
+  // initialize Wire Protocol semaphore in the "not taken" state
+  chBSemObjectInit(&wpChannelSemaphore, false);
 
   // loop until thread receives a request to terminate
   while (!chThdShouldTerminateX()) {
